requirements.txt
pandas==0.23.4
matplotlib==2.2.3
jupyter
box2d-py==2.3.4
gym==0.10.5
unityagents==0.4.0
h5py==2.8.0
tqdm==4.25.0
scikit-image==0.14.0
nose==1.3.7
nose-parameterized==0.6.0
easydict==1.8
torch==0.4.1
torchvision==0.2.1